Simplified gamma function; generalized parameters
- gamma function had strange behavior when running
under Valgrind, so I simplified it a bit
- n and l can now be double, not just int, useful for
integrals of ordinary Bessel functions

+++ src/integrals.c
@@ -463,8 +463,8 @@ int coffe_integrals_renormalizable(
     double *output_y,
     const size_t output_len,
     const struct coffe_interpolation *spectrum,
-    const int l,
-    const int n,
+    const double l,
+    const double n,
     const double x_min,
     const double x_max
 )
